Update to handle a 401 from the API
Registeres a global jQuery AJAX error handler that will unset the token and
attempt to authenticate with GitHub again.

Move the `Backbone.history` initialization into the `init` function so that it
will only ever be called once.

Adds an `auth` function that can be called in the AJAX error handler.

Closes #5