Fix issue with events becoming unbound
The view for NotificationDetails was detached after a timeout, but when
another view replaced the contents of the #details container, it would
called `$(“#details”).html(…)`, which would remove the element before
it was detached, unbinding any events in the process. This works around
it by appending all loaded views and using CSS to determine which one
to show.