Explicitly declare when to bind the space keyboard event
…instead of relying on `model.isSelected()`, which doesn’t always work
here since it only deals with the model’s primary collection and not
the selected state within the `Timeline` collection. For example, an
event gets selected by calling `@timeline.select event`, which makes it
selected in the timeline collection, but `model.isSelected()` returns
false since the event is not selected in the original `Events`
collection. This is probably considered a bug in the way that the
selected state is managed for models in multiple collections, but that
will be a lot more complicated to fix.